Some little modifications

+++ src/main/java/com/example/idcard/rest/PersonController.java
@@ -4,6 +4,7 @@ import com.example.idcard.dto.PersonDto;
 import com.example.idcard.model.Person;
 import com.example.idcard.repository.PersonRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -26,6 +27,14 @@ public class PersonController {
         return personRepo.getById(id);
     }
 
+    @PostMapping("/update/{id}")
+    public void update(@RequestBody Person person,int id){
+        Person person1 = null;
+        person1 = personRepo.findById(id);
+        person1 = person;
+
+        personRepo.save(person1);
+    }
 
 
     @PostMapping("/create")
@@ -33,9 +42,14 @@ public class PersonController {
         personRepo.save(person);
     }
 
-    @PostMapping("/delete/{id}")
-    public void delete(@PathVariable int id){
-        personRepo.deleteById(id);
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity delete(@PathVariable int id){
+        if(personRepo.existsById(id)){
+            personRepo.deleteById(id);
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body("person with id: "+id+" is deleted");
+        }else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("There is no person with this id");
+        }
     }
